requirements.txt
asgiref==3.5.1
Django==4.0.4
django-admin-rangefilter==0.8.4
sqlparse==0.4.2
python-dotenv
